+++ src/lib.rs
@@ -3,6 +3,7 @@ pub mod colors;
 pub mod input;
 pub mod particles;
 pub mod rapier;
+pub mod winner;
 
 pub mod prelude {
     pub use crate::audio::*;
@@ -10,4 +11,5 @@ pub mod prelude {
     pub use crate::input::*;
     pub use crate::particles::*;
     pub use crate::rapier::*;
+    pub use crate::winner::*;
 }